requirements.txt
streamlit
pillow
python-dotenv
inference_sdk